refactor: move chat engine from vectordbs to engine folder
