+++ templates/components/engines/typescript/agent/tools.ts
@@ -1,0 +1,1 @@
+// Load tools here
