+++ templates/components/vectordbs/typescript/none/index.ts
@@ -1,5 +1,4 @@
 import {
-  ContextChatEngine,
   LLM,
   serviceContextFromDefaults,
   SimpleDocumentStore,
@@ -8,7 +7,7 @@ import {
 } from "llamaindex";
 import { CHUNK_OVERLAP, CHUNK_SIZE, STORAGE_CACHE_DIR } from "./constants.mjs";
 
-async function getDataSource(llm: LLM) {
+export async function getDataSource(llm: LLM) {
   const serviceContext = serviceContextFromDefaults({
     llm,
     chunkSize: CHUNK_SIZE,
@@ -29,16 +28,5 @@ async function getDataSource(llm: LLM) {
   return await VectorStoreIndex.init({
     storageContext,
     serviceContext,
-  });
-}
-
-export async function createChatEngine(llm: LLM) {
-  const index = await getDataSource(llm);
-  const retriever = index.asRetriever();
-  retriever.similarityTopK = 3;
-
-  return new ContextChatEngine({
-    chatModel: llm,
-    retriever,
   });
 }
